backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
distlib==0.3.8
Django==4.2.13
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
filelock==3.14.0
pipenv==2023.12.1
platformdirs==4.2.2
sqlparse==0.5.0
typing_extensions==4.12.1
virtualenv==20.26.2
